Move claim-related methods out of the datastore class to a manager class
